check database data and correct some variables

+++ src/main/java/com/abseliamov/cinema/controller/MovieController.java
@@ -3,6 +3,8 @@ package com.abseliamov.cinema.controller;
 import com.abseliamov.cinema.model.Movie;
 import com.abseliamov.cinema.service.MovieService;
 
+import java.math.BigDecimal;
+
 public class MovieController {
     private MovieService movieService;
 
@@ -10,11 +12,11 @@ public class MovieController {
         this.movieService = movieService;
     }
 
-    public boolean increaseCostMovie(double ticketCost, Movie movie) {
+    public boolean increaseCostMovie(BigDecimal ticketCost, Movie movie) {
         return movieService.increaseCostMovie(ticketCost, movie);
     }
 
-    public boolean reduceCostMovie(double ticketCost, Movie movie) {
+    public boolean reduceCostMovie(BigDecimal ticketCost, Movie movie) {
         return movieService.reduceCostMovie(ticketCost, movie);
     }
